requirements.txt
canonicalwebteam.flask-base==1.0.5
google-api-python-client==1.8.2
django-openid-auth==0.16
Flask-OpenID==1.3.0
beautifulsoup4==4.11.1
lxml==4.8.0
python-dateutil==2.8.2
# TODO: remove this library once we have access to the scope drive.readonly
google-auth-oauthlib==0.5.1
